+++ prisma/migrations/20240406201916_user_address/migration.sql
@@ -1,0 +1,20 @@
+-- CreateTable
+CREATE TABLE "UserAddress" (
+    "id" TEXT NOT NULL,
+    "firstName" TEXT NOT NULL,
+    "lastName" TEXT NOT NULL,
+    "address" TEXT NOT NULL,
+    "address2" TEXT,
+    "podtalCode" TEXT NOT NULL,
+    "telephone" TEXT NOT NULL,
+    "idUser" TEXT NOT NULL,
+    "countryId" TEXT NOT NULL,
+
+    CONSTRAINT "UserAddress_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "UserAddress" ADD CONSTRAINT "UserAddress_idUser_fkey" FOREIGN KEY ("idUser") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "UserAddress" ADD CONSTRAINT "UserAddress_countryId_fkey" FOREIGN KEY ("countryId") REFERENCES "Country"("id") ON DELETE RESTRICT ON UPDATE CASCADE;