+++ src/actions/index.ts
@@ -5,3 +5,14 @@ export * from './products/get-stock-by-slug'
 export * from './auth/login'
 export * from './auth/logout'
 export * from './auth/registerUser'
+
+export * from './country/get-countries'
+
+export * from './order/place-order'
+export * from './order/get-order-by-id'
+export * from './order/get-orders-by-user'
+export * from './order/orders-pagination'
+
+export * from './address/set-user-address'
+export * from './address/delete-user-address'
+export * from './address/get-user-address'