+++ src/app/(shop)/profile/page.tsx
@@ -1,14 +1,8 @@
 import { auth } from '@/auth.config'
 import { Title } from '@/components'
-import { redirect } from 'next/navigation'
 
 export default async function ProfilePage() {
   const session = await auth()
-
-  // if (!session?.user) {
-  //   // redirect('/auth/login?returnTo=perfil')
-  //   redirect('/')
-  // }
 
   return (
     <>