+++ src/components/product/quantity-selector/QuantitySelector.tsx
@@ -13,7 +13,6 @@ export const QuantitySelector = ({ quantity, stock, onSetQuantity }: Props) => {
   const quantityChange = (value: number) => {
     const newQuantity = quantity + value
     if (newQuantity > stock || newQuantity < 1) return
-    console.log('pase el if', newQuantity, stock)
     onSetQuantity(newQuantity)
   }