+++ src/validations/index.ts
@@ -1,1 +1,2 @@
 export * from './auth/new-account.validations'
+export * from './address/address.validations'